+++ HISTORY.rst
@@ -8,6 +8,8 @@
 **Enhancements**
 
 - 971_: [Linux] Add psutil.sensors_fans() function.  (patch by Nicolas Hennion)
+- 976_: [Windows] Process.io_counters() has 2 new fields: *other_count* and
+  *other_bytes*.
 
 5.1.3
 =====